Fix IC-3 validation query.
Issue #1 - Update the IC-3 validation query to handle components which
directly declare a qb:measure property. The IC-* validation queries
in the specification are defined against a 'normalised' cube (see
section 10.1 of the data cube specification). The normalisation
process inserts an extra qb:componentProperty statement for components
which directly declare a qb:meaasure property. The existing validation
query assumes this statement exists when it may not.

The updated query searches for DSD components of two types:
1. Those that declare a qb:measure directly.
2. Those that declare a qb:componentProperty (or sub-property) of type
   qb:MeasureType.

The query returns DSDs which have no components which satisfy either
of the two cases.